Try to work with remote wishList
