+++ src/components/UserWishList/UserWishList.jsx
@@ -1,6 +1,5 @@
 import { useState, useEffect } from 'react';
 import { useSelector, connect } from 'react-redux';
-import axios from 'axios';
 import { instance } from '../../components/assets/axiosUrl';
 import { Link } from 'react-router-dom';
 
@@ -33,44 +32,44 @@ const UserWishList = ({ token }) => {
     // }, [token]);
 
     // Get wishlist for User - is working
-    useEffect(() => {
-        async function fetchWishListItems () {
-            try {
-                const response = await instance.get('/api/wishlist', {
-                    headers: { Authorization: token }
-                });
-                const wishlist = response.data;
-                console.log('Wishlist from UserWishList', wishlist);
-                setWishListItems(wishlist);
-                setWishListItemIsLoading(false);
-            } catch (err) {
-                console.log('Error from get UserWishList', err);
-                setWishListItemIsLoading(false);
-            }
-        }
+    // useEffect(() => {
+    //     async function fetchWishListItems () {
+    //         try {
+    //             const response = await instance.get('/api/wishlist', {
+    //                 headers: { Authorization: token }
+    //             });
+    //             const wishlist = response.data;
+    //             console.log('Wishlist from UserWishList', wishlist);
+    //             setWishListItems(wishlist);
+    //             setWishListItemIsLoading(false);
+    //         } catch (err) {
+    //             console.log('Error from get UserWishList', err);
+    //             setWishListItemIsLoading(false);
+    //         }
+    //     }
 
-        fetchWishListItems();
-    }, [token]);
+    //     fetchWishListItems();
+    // }, [token]);
 
     // Add product to Wishlist for User - is working
-    async function addProductToWishList (productId) {
-        console.log(productId);
-        if (wishListItemIsLoading) {
-            if (!wishListItems || !wishListItems.products.some(item => item._id === productId)) {
-                try {
-                    const response = await instance.put(`/api/wishlist/${productId}`, null, {
-                        headers: { Authorization: token }
-                    });
-                    const updatedWishlist = response.data;
-                    console.log('UpdatedWishlist from UserWishList', updatedWishlist);
-                    setWishListItems(updatedWishlist);
-                } catch (err) {
-                    console.log('Error from ADD UserWishList', err);
-                }
-            }
-        }
-    }
-    addProductToWishList('64fc0ca8b67efd9a68af20bb')
+    // async function addProductToWishList (productId) {
+    //     console.log(productId);
+    //     if (wishListItemIsLoading) {
+    //         if (!wishListItems || !wishListItems.products.some(item => item._id === productId)) {
+    //             try {
+    //                 const response = await instance.put(`/api/wishlist/${productId}`, null, {
+    //                     headers: { Authorization: token }
+    //                 });
+    //                 const updatedWishlist = response.data;
+    //                 console.log('UpdatedWishlist from UserWishList', updatedWishlist);
+    //                 setWishListItems(updatedWishlist);
+    //             } catch (err) {
+    //                 console.log('Error from ADD UserWishList', err);
+    //             }
+    //         }
+    //     }
+    // }
+    // addProductToWishList('64fc0ca8b67efd9a68af20bb')
 
     return (
         <div className={style.userWishList}>
